% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{create_tip_names}
\alias{create_tip_names}
\title{Create tip names}
\usage{
create_tip_names(region_n = 100, exog_n = 50)
}
\arguments{
\item{region_n}{numeric which is defined as the number of tips for the region
data. Region data are data for the region of interest. Default is set to 100.}

\item{exog_n}{numeric which is defined as the number of tips for exogenous
data. Exogenous data are data for other regions that are not the region of
interest. Default is set to 50.}
}
\value{
dataframe for tip names and states associated to each tip name
}
\description{
This function will create names for terminal branches (tip) in the
phylogenetic tree. It will also associate each tip to a deme from the
structured coalescent model.
}
\examples{
tip_data <- create_tip_names()
}
