% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{createSample_times}
\alias{createSample_times}
\title{Create time for terminal branches in the phylogenetic tree}
\usage{
createSample_times(time_region, time_exog, state_info)
}
\arguments{
\item{time_region}{numeric which is the start time for sampling}

\item{time_exog}{numeric which is the maximum time for sampling}

\item{state_info}{dataframe containing the information of tip names and state.
State is a character associating each tip to a deme of the structured
coalescent model.
#'}
}
\value{
numeric vector for sampling times. Each value in the vector is
named using the name used for the tips in the phylogenetic tree.
}
\description{
Function to sample time for each terminal branch in a phylogenetic tree.
This sampling will be stratified through time, which is equal time increments
between samples. Samples here is each terminal branch in the phylogenetic tree.
}
\examples{
tip_data <- create_tip_names()
sampleTimes <- createSample_times(time_region = 2020.15,
                                  time_exog = 2020.0,
                                  state_info = tip_data)
}
