% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{rename_tips}
\alias{rename_tips}
\title{Rename tip names}
\usage{
rename_tips(tip_name, tip_time)
}
\arguments{
\item{tip_name}{tip name after spliting the name into two: first part contains
the tip name, second part contains the state name. For example: if tip name is
"Seq1_exog", the splitted form will be "Seq1" "exog".}

\item{tip_time}{time for correspondent tip name}
}
\value{
new tip name
}
\description{
Function to rename tip names to include information of sample times
}
\examples{
TO DO
}
