% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sampler}
\alias{sampler}
\title{Sampler to sample parameters from a uniform distribution}
\usage{
sampler(n = 100)
}
\arguments{
\item{n}{numeric for the number of samples. Default is set to 100.}
}
\value{
dataframe for parameters sampled.
}
\description{
Funtion that will sample parameter values from a uniform distribution based
on latin hypercube sampling.
}
\details{
This function will sample from a uniform distribution parameter
 values for b, initS (initial size of susceptibles), import (importation rate),
  st (start time) and E (initial size of exposed individuals).
 Boundaries for each parameter:
 b: min = 15, max = 25
 initS: min = 1e4, max = 1e6
 import: min = 1, max = 10
 st: min = 2020.10, max = 2020.15
 E: min = 1, max = 30
}
\examples{
#the example below will sample 100 parameter values for each b, initS, import,
and st from a uniform ditribution
parameters <- sampler()
}
