% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_dna}
\alias{simulate_dna}
\title{Simulate DNA sequence alignment}
\usage{
simulate_dna(tree)
}
\arguments{
\item{tree}{object of class phylo}
}
\value{
DNA sequence alignment in PHYLIP format
}
\description{
Function to simulate DNA sequence alignment based on the HKY DNA substitution
model
}
\details{
This function will simulate DNA sequence alignment using seq-gen
\url{http://tree.bio.ed.ac.uk/software/seqgen/}.
Parameters used in this function are:
-mHKY: selects the model. Here we select HKY DNA substitution model
-l29500: selects the length of the DNA sequence alignment to simulate. Here
  alignments will have 29500 base pairs.
-t5.5: selects the transition/tranversion rate
-f0.3,0.2,0.2,0.3 selects the frequency of each nucleotide in the sequence
  alignment in the order of A, C, G, T.
 Parameters -t and -f must be set when simulating HKY substitution models
-s0.001 scale branch lengths in order to make then equal to the expected
  number of substitution per site for each branch.
-or: Return alignment as a relaxed PHYLIP format. With this option sequence
  names can have more than 10 characters.
}
\examples{
TO DO
}
