% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{write_ali}
\alias{write_ali}
\title{Save DNA sequence alignment in a file}
\usage{
write_ali(seqgen_ali, filename)
}
\arguments{
\item{seqgen_ali}{object of class seqgen}

\item{filename}{name for the file to save sequence alignment. Note that
this name should NOT have the extention .phy, for example.}
}
\value{

}
\description{
Funtion to save seq-gen DNA sequence alignment into a file.
}
\examples{
TO DO
}
